mui-dashboard/requirements.txt
Flask
flask-cors
openpyxl
pandas
google-generativeai
google-api-python-client
math
re
json
os
logging
